hippocampal_dynamics_toolbox/requirements.txt
annotated-types==0.7.0
asciitree==0.3.3
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.2.18
fasteners==0.20
fonttools==4.60.1
greenlet==3.2.4
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
neo==0.14.3
numcodecs==0.15.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.55.0
probeinterface==0.3.0
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
quantities==0.16.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
spikeinterface==0.103.0
tensorpac==0.6.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
wrapt==1.17.3
zarr==2.18.7
